backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg[binary]
alembic
celery
redis
httpx
pytest
